requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.2
orjson==3.10.0
pandas==2.1.4
pydantic==2.5.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.32.0.post1
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
xgboost==2.0.2
